Created MenuLinkTrait so we can share menu links for Functional tests as well
